requirements.txt
Django==2.2.6
django-debug-toolbar==3.2
Pillow==8.2.0
pytz==2021.1
sorl-thumbnail==12.6.3
sqlparse==0.4.1
